+++ jni/zip_source_filep.c
@@ -140,7 +140,7 @@ read_file(void *state, void *data, size_t len, enum zip_source_cmd cmd)
 	
     case ZIP_SOURCE_READ:
 	if (z->remain != -1)
-	    n = len > z->remain ? z->remain : len;
+	    n = len > z->remain ? (int)z->remain : (int)len;
 	else
 	    n = len;
